requirements.txt
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==40.1.2
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
packaging==25.0
pillow==12.1.0
pycparser==2.23
pydyf==0.12.1
PyJWT==2.10.1
pyparsing==3.3.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.3
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
